requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.9
django-filter==23.5
djangorestframework==3.15.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
